Reorganized the command/cog files into proper cog directories
- Renamed cog/command files to have the `_cog` suffix
- Tweaked imports to play nice with new paths
