+++ code/common/cogs/invite_cog.py
@@ -16,7 +16,7 @@ CONFIG_OPTIONS = Configuration.load_config()
 LOGGER = Logging.initialize_logging(logging.getLogger(__name__))
 
 
-class InviteCommand(Cog):
+class InviteCog(Cog):
 
     def __init__(self, bot: Bot, *args, **kwargs):
         super().__init__(bot, *args, **kwargs)
