+++ code/core/cogs/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    'admin',
+    'help_cog',
+    'speech_config_'
+]
