--- code/core/commands/__init__.py
@@ -1,6 +1,0 @@
-__all__ = [
-    'admin',
-    'help_command',
-    'invite_command',
-    'speech_config_help_command'
-]
