+++ code/hawking.py
@@ -13,14 +13,10 @@ import asyncio
 import discord
 from discord.ext import commands
 
-from core import speech
-from core import message_parser
-from core.commands import admin
-from core.commands import help_cog
-from core.commands import invite_command
-from core.commands import speech_config_help_command
+from core import speech, message_parser
+from core.cogs import admin_cog, help_cog, speech_config_help_cog
 from common import audio_player
-from common import privacy_manager
+from common.cogs import privacy_management_cog, invite_cog
 from common.configuration import Configuration
 from common.logging import Logging
 from common.command_management import invoked_command_handler, command_reconstructor
@@ -83,23 +79,23 @@ class Hawking:
             dependencies=[database_manager.DatabaseManager]
         )
         self.module_manager.register_module(
-            admin.Admin,
+            admin_cog.AdminCog,
             self,
             self.bot,
             dependencies=[database_manager.DatabaseManager]
         )
         self.module_manager.register_module(
-            privacy_manager.PrivacyManager,
+            privacy_management_cog.PrivacyManagementCog,
             self.bot,
             dependencies=[component_factory.ComponentFactory, database_manager.DatabaseManager]
         )
         self.module_manager.register_module(
-            speech_config_help_command.SpeechConfigHelpCommand,
+            speech_config_help_cog.SpeechConfigHelpCog,
             self.bot,
             dependencies=[component_factory.ComponentFactory, database_manager.DatabaseManager]
         )
         self.module_manager.register_module(
-            invite_command.InviteCommand,
+            invite_cog.InviteCog,
             self.bot,
             dependencies=[component_factory.ComponentFactory, database_manager.DatabaseManager]
         )
@@ -110,7 +106,7 @@ class Hawking:
         self.module_manager.register_module(
             audio_player.AudioPlayer,
             self.bot,
-            dependencies=[admin.Admin, database_manager.DatabaseManager]
+            dependencies=[admin_cog.AdminCog, database_manager.DatabaseManager]
         )
         self.module_manager.register_module(
             speech.Speech,
